Clean legacy code and move into legacy folder
Remove unused private method, remove analysis scripts and move remaining files to legacy code folder.
