+++ extract_cxd.py
@@ -6,6 +6,7 @@ import csv
 import math
 import struct
 import olefile # pip install olefile
+import numpy as np
 
 
 class ExtractError(Exception):
@@ -61,7 +62,7 @@ def read_debug(ole, stream):
     print("%s: %s%s" % (stream, ":".join("{:02x}".format(ord(c)) for c in data), a))
 
 
-def _extract_cxd(ole, to_dir, prefix):
+def _extract_cxd(ole, to_file):
     if not ole.exists('File Info/Field Count'):
         raise ExtractError('Not an HCImage container')
 
@@ -74,39 +75,56 @@ def _extract_cxd(ole, to_dir, prefix):
     if ole.exists('File Data/Field 1/i_image1/Bitmap 2'):
         raise ExtractError('Unsupported: multiple bitmaps per image')
 
-    # digits
-    digits = 1 + int(math.floor(math.log10(expected_frames)))
+    # expected properties based on first frame
+    height = int(read_double(ole, 'Field Data/Field 1/i_Image1/Details/Image_Height'))
+    width = int(read_double(ole, 'Field Data/Field 1/i_Image1/Details/Image_Width'))
+    depth = int(read_double(ole, 'Field Data/Field 1/i_Image1/Details/Image_Depth'))
+    bytes_per_pixel = int(math.ceil(depth / 8.))
+    bytes_per_frame = width * height * bytes_per_pixel
 
-    # for each frame
-    rows = []
-    for frame in xrange(1, expected_frames + 1):
-        # frame details
-        time_from_last = read_double(ole, 'Field Data/Field %d/Details/Time_From_Last' % frame)
-        time_from_start = read_double(ole, 'Field Data/Field %d/Details/Time_From_Start' % frame)
-        exposure = read_double(ole, 'Field Data/Field %d/i_Image1/Details/Image_Exposure1' % frame)
+    # numpy formats
+    dtype_read = np.dtype('u%d' % bytes_per_pixel).newbyteorder('<')
+    dtype_write = np.dtype('u%d' % bytes_per_pixel)
+    # video = np.ndarray((height, width, expected_frames), dtype=np.dtype('u%d' % bytes_per_pixel))
 
-        # image details
-        binning = int(read_double(ole, 'Field Data/Field %d/i_Image1/Details/Binning' % frame))
-        depth = int(read_double(ole, 'Field Data/Field %d/i_Image1/Details/Image_Depth' % frame))
-        height = int(read_double(ole, 'Field Data/Field %d/i_Image1/Details/Image_Height' % frame))
-        width = int(read_double(ole, 'Field Data/Field %d/i_Image1/Details/Image_Width' % frame))
+    # open video file
+    fn = to_file + '.video'
+    with open(fn, 'wb') as f:
+        # for each frame
+        rows = []
+        for frame in range(1, expected_frames + 1):
+            # frame details
+            time_from_last = read_double(ole, 'Field Data/Field %d/Details/Time_From_Last' % frame)
+            time_from_start = read_double(ole, 'Field Data/Field %d/Details/Time_From_Start' % frame)
+            cur_exposure = read_double(ole, 'Field Data/Field %d/i_Image1/Details/Image_Exposure1' % frame)
 
-        # load image
-        expected_bytes = width * height * int(math.ceil(depth / 8.))
-        bitmap = read_data(ole, 'Field Data/Field %d/i_Image1/Bitmap 1' % frame, expected_length=expected_bytes)
+            # image details
+            cur_binning = int(read_double(ole, 'Field Data/Field %d/i_Image1/Details/Binning' % frame))
+            cur_depth = int(read_double(ole, 'Field Data/Field %d/i_Image1/Details/Image_Depth' % frame))
+            cur_height = int(read_double(ole, 'Field Data/Field %d/i_Image1/Details/Image_Height' % frame))
+            cur_width = int(read_double(ole, 'Field Data/Field %d/i_Image1/Details/Image_Width' % frame))
 
-        # write bitmap
-        fn = '%%0%dd' % digits % frame
-        fn = os.path.join(to_dir, prefix + fn + '.frame')
-        with open(fn, 'wb') as f:
-            f.write(bitmap)
+            # check expectations
+            if cur_depth != depth or cur_height != height or cur_width != width:
+                raise ExtractError('Unsupported: change in depth (%d/%d), width (%d/%d) or height (%d/%d)'
+                                   % (depth, cur_depth, width, cur_width, height, cur_height))
 
-        # append row
-        rows.append((frame, binning, depth, height, width, time_from_last, exposure))
+            # load bitmap
+            bitmap_data = read_data(ole, 'Field Data/Field %d/i_Image1/Bitmap 1' % frame, expected_length=bytes_per_frame)
+            bitmap_array = np.fromstring(bitmap_data, dtype=dtype_read)
+
+            # write bytes
+            bitmap_array.astype(dtype_write).tofile(f)
+
+            # decode
+            #video[:, :, frame - 1] = bitmap_array.reshape((height, width))
+
+            # append row
+            rows.append((frame, cur_binning, cur_depth, cur_height, cur_width, time_from_last, cur_exposure))
 
     # write csv
-    fn = os.path.join(to_dir, prefix + 'frames.csv')
-    with open(fn, 'wb') as f:
+    fn = to_file + '.csv'
+    with open(fn, 'w') as f:
         writer = csv.writer(f, doublequote=False, escapechar='\\', lineterminator='\n')
         for row in rows:
             writer.writerow([str(x) for x in row])
@@ -114,7 +132,7 @@ def _extract_cxd(ole, to_dir, prefix):
     return expected_frames
 
 
-def extract_cxd(from_cxd, to_dir, prefix=None):
+def extract_cxd(from_cxd, to_file):
     if not olefile.isOleFile(from_cxd):
         raise ExtractError('Not an OLE container')
 
@@ -122,7 +140,7 @@ def extract_cxd(from_cxd, to_dir, prefix=None):
     ole = olefile.OleFileIO(from_cxd)
 
     try:
-        return _extract_cxd(ole, to_dir, '' if prefix is None else prefix)
+        return _extract_cxd(ole, to_file)
     finally:
         ole.close()
 
@@ -132,33 +150,37 @@ def eprint(*args, **kwargs):
 
 
 def main():
-    if len(sys.argv) != 3:
+    if len(sys.argv) != 2 and len(sys.argv) != 3:
         print('Usage:')
         print('')
-        print('\t%s input-cxd output-dir' % sys.argv[0])
+        print('\t%s input-cxd [output-file]' % sys.argv[0])
         print('')
         sys.exit(0 if len(sys.argv) == 1 else 1)
 
     # read arguments
     in_cxd = sys.argv[1]
-    out_dir = sys.argv[2]
+    out_file = in_cxd if len(sys.argv) < 3 else sys.argv[2]
 
     # check for file existence
     if not os.path.isfile(in_cxd):
         eprint('File not found: %s' % in_cxd)
         sys.exit(1)
 
-    if not os.path.isdir(out_dir):
-        eprint('Directory not found: %s' % out_dir)
+    if not os.path.isdir(os.path.dirname(out_file)):
+        eprint('Directory not found: %s' % os.path.dirname(out_file))
         sys.exit(1)
 
-    # figure out prefix from file name
-    name, _ = os.path.splitext(os.path.basename(in_cxd))
-    prefix = name + '_'
+    # strip extension?
+    if '.' in os.path.basename(out_file):
+        out_file = os.path.splitext(out_file)[0]
+
+    if os.path.isfile(out_file):
+        eprint('Output already exists: %s' % out_file)
+        sys.exit(1)
 
     # extract cxd file
     try:
-        extracted = extract_cxd(in_cxd, out_dir, prefix)
+        extracted = extract_cxd(in_cxd, out_file)
     except ExtractError as e:
         eprint('Error: %s' % e)
         sys.exit(1)
