+++ legacy/match_template.m
