+++ legacy/private/plot_bci.m
