+++ legacy/private/regions_of_true.m
