+++ legacy/private/video_smooth2.m
