+++ legacy/private/video_write.m
