--- video_crop_file.m
@@ -1,8 +1,0 @@
-function struct = video_crop_file(struct,  params)
-%VIDEO_CROP_FILE Summary of this function goes here
-%   Detailed explanation goes here
-
-struct.video = video_crop(struct.video, 'objective', params.objective);
-
-end
-
